requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==43.0.0
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
html5lib==1.1
idna==3.6
M2Crypto==0.41.0
numpy==1.24.4
pandas==2.0.3
pillow==10.2.0
ply==3.11
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.9.0
pyphen==0.14.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing-extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==6.6.0
